refactor task usecases
