--- src/main/kotlin/logic/usecase/state/CheckTaskStateExistsUseCase.kt
@@ -1,11 +1,0 @@
-package logic.usecase.state
-
-import logic.repo.TaskStateRepository
-
-class CheckTaskStateExistsUseCase(
-    private val repository: TaskStateRepository
-) {
-    operator fun invoke(name: String): Boolean {
-        return repository.exist(name)
-    }
-}
