--- src/main/kotlin/logic/usecase/task/CreateTaskUseCase.kt
@@ -1,8 +1,0 @@
-package logic.usecase.task
-
-import logic.model.Task
-import logic.repo.TaskRepository
-
-class CreateTaskUseCase(private val repository: TaskRepository) {
-    operator fun invoke(task: Task) = repository.create(task)
-}
