--- src/main/kotlin/logic/usecase/task/DeleteTaskUseCase.kt
@@ -1,8 +1,0 @@
-package logic.usecase.task
-
-import logic.repo.TaskRepository
-import java.util.*
-
-class DeleteTaskUseCase(private val repository: TaskRepository) {
-    operator fun invoke(taskId: UUID) = repository.delete(taskId)
-}
