--- src/main/kotlin/logic/usecase/task/GetTasksByProjectUseCase.kt
@@ -1,9 +1,0 @@
-package logic.usecase.task
-
-import logic.model.Task
-import logic.repo.TaskRepository
-import java.util.*
-
-class GetTasksByProjectUseCase(private val repository: TaskRepository) {
-    operator fun invoke(projectId: UUID): List<Task> = repository.getByProject(projectId)
-}
